NF: move UndoManager to its file
