+++ AnkiDroid/src/main/java/com/ichi2/libanki/Collection.java
@@ -20,7 +20,6 @@ package com.ichi2.libanki;
 import android.annotation.SuppressLint;
 import android.content.ContentValues;
 import android.content.Context;
-import android.content.res.Resources;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabaseLockedException;
 import android.text.TextUtils;
@@ -72,7 +71,6 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Random;
-import java.util.concurrent.LinkedBlockingDeque;
 import java.util.Set;
 import java.util.regex.Pattern;
 
@@ -1291,96 +1289,6 @@ public class Collection implements CollectionGetter {
             return new Pair<>(mConf.getInt("timeLim"), mSched.getReps() - mStartReps);
         }
         return null;
-    }
-
-
-    /**
-     * Undo ********************************************************************* **************************
-     */
-
-    public static class UndoManager {
-        private LinkedBlockingDeque<Undoable> mUndo;
-        private static final int UNDO_SIZE_MAX = 20;
-
-
-        // API 21: Use a ConcurrentLinkedDeque
-        /* Note from upstream:
-         * this data structure is a mess, and will be updated soon
-         * in the review case, [1, "Review", [firstReviewedCard, secondReviewedCard, ...], wasLeech]
-         * in the checkpoint case, [2, "action name"]
-         * wasLeech should have been recorded for each card, not globally
-         */
-        public void clearUndo() {
-            mUndo = new LinkedBlockingDeque<>();
-        }
-
-
-        public void markUndo(@NonNull Undoable undo) {
-            Timber.d("markUndo() of type %s", undo.getClass());
-            this.mUndo.add(undo);
-            while (this.mUndo.size() > UNDO_SIZE_MAX) {
-                this.mUndo.removeFirst();
-            }
-        }
-
-
-        /** Undo menu item name, or "" if undo unavailable.*/
-        @VisibleForTesting
-        public @Nullable Undoable undoType() {
-            if (mUndo.size() > 0) {
-                return mUndo.getLast();
-            }
-            return null;
-        }
-
-
-        public boolean undoAvailable() {
-            Timber.d("undoAvailable() undo size: %s", mUndo.size());
-            return mUndo.size() > 0;
-        }
-
-
-        public @Nullable Card undo(Collection collection) {
-            Undoable lastUndo = mUndo.removeLast();
-            Timber.d("undo() of type %s", lastUndo.getClass());
-            return lastUndo.undo(collection);
-        }
-
-
-        public String undoName(Resources res) {
-            Undoable type = undoType();
-            if (type != null) {
-                return type.name(res);
-            }
-            return "";
-        }
-
-
-        public void markReview(Card card, Collection collection) {
-            boolean wasLeech = card.note().hasTag("leech");
-            Card clonedCard = card.clone();
-            markUndo(new UndoReview(wasLeech, clonedCard));
-        }
-    }
-
-
-
-    @VisibleForTesting
-    public static class UndoReview extends Undoable {
-        private final boolean mWasLeech;
-        @NonNull private final Card mClonedCard;
-        public UndoReview(boolean wasLeech, @NonNull Card clonedCard) {
-            super(R.string.undo_action_review);
-            mClonedCard = clonedCard;
-            mWasLeech = wasLeech;
-        }
-
-        @NonNull
-        @Override
-        public Card undo(@NonNull Collection col) {
-            col.getSched().undoReview(mClonedCard, mWasLeech);
-            return mClonedCard;
-        }
     }
